refactor(multi-select): rewrite sorting in typescript
